Moved CredentialsProvider to a separate file, added type hints
