+++ redis/connection.py
@@ -11,6 +11,7 @@ from time import time
 from urllib.parse import parse_qs, unquote, urlparse
 
 from redis.backoff import NoBackoff
+from redis.credentials import CredentialsProvider
 from redis.exceptions import (
     AuthenticationError,
     AuthenticationWrongNumberOfArgsError,
@@ -474,42 +475,6 @@ if HIREDIS_AVAILABLE:
     DefaultParser = HiredisParser
 else:
     DefaultParser = PythonParser
-
-
-class CredentialsProvider:
-    def __init__(self, username="", password="", supplier=None, *args, **kwargs):
-        """
-        Initialize a new Credentials Provider.
-        :param supplier: a supplier function that returns the username and password.
-                         def supplier(arg1, arg2, ...) -> (username, password)
-                         For examples see examples/connection_examples.ipynb
-        :param args: arguments to pass to the supplier function
-        :param kwargs: keyword arguments to pass to the supplier function
-        """
-        self.username = username
-        self.password = password
-        self.supplier = supplier
-        self.args = args
-        self.kwargs = kwargs
-
-    def get_credentials(self):
-        if self.supplier:
-            self.username, self.password = self.supplier(*self.args, **self.kwargs)
-        if self.username:
-            auth_args = (self.username, self.password or "")
-        else:
-            auth_args = (self.password,)
-        return auth_args
-
-    def get_password(self, call_supplier=True):
-        if call_supplier and self.supplier:
-            self.username, self.password = self.supplier(*self.args, **self.kwargs)
-        return self.password
-
-    def get_username(self, call_supplier=True):
-        if call_supplier and self.supplier:
-            self.username, self.password = self.supplier(*self.args, **self.kwargs)
-        return self.username
 
 
 class Connection:
