+++ tests/test_connection.py
@@ -1,25 +1,17 @@
-import random
 import socket
-import string
 import types
 from unittest import mock
 from unittest.mock import patch
 
 import pytest
 
-import redis
 from redis.backoff import NoBackoff
-from redis.connection import Connection, CredentialsProvider
-from redis.exceptions import (
-    ConnectionError,
-    InvalidResponse,
-    ResponseError,
-    TimeoutError,
-)
+from redis.connection import Connection
+from redis.exceptions import ConnectionError, InvalidResponse, TimeoutError
 from redis.retry import Retry
 from redis.utils import HIREDIS_AVAILABLE
 
-from .conftest import _get_client, skip_if_redis_enterprise, skip_if_server_version_lt
+from .conftest import skip_if_server_version_lt
 
 
 @pytest.mark.skipif(HIREDIS_AVAILABLE, reason="PythonParser only")
@@ -130,86 +122,3 @@ class TestConnection:
         assert conn._connect.call_count == 1
         assert str(e.value) == "Timeout connecting to server"
         self.clear(conn)
-
-
-class TestCredentialsProvider:
-    @skip_if_redis_enterprise()
-    def test_credentials_provider_without_supplier(self, r, request):
-        # first, test for default user (`username` is supposed to be optional)
-        default_username = "default"
-        temp_pass = "temp_pass"
-        creds_provider = CredentialsProvider(default_username, temp_pass)
-        r.config_set("requirepass", temp_pass)
-        creds = creds_provider.get_credentials()
-        assert r.auth(creds[1], creds[0]) is True
-        assert r.auth(creds_provider.get_password()) is True
-
-        # test for other users
-        username = "redis-py-auth"
-        password = "strong_password"
-
-        def teardown():
-            try:
-                r.auth(temp_pass)
-            except ResponseError:
-                r.auth("default", "")
-            r.config_set("requirepass", "")
-            r.acl_deluser(username)
-
-        request.addfinalizer(teardown)
-
-        assert r.acl_setuser(
-            username,
-            enabled=True,
-            passwords=["+" + password],
-            keys="~*",
-            commands=["+ping", "+command", "+info", "+select", "+flushdb", "+cluster"],
-        )
-
-        creds_provider2 = CredentialsProvider(username, password)
-        r2 = _get_client(
-            redis.Redis, request, flushdb=False, credentials_provider=creds_provider2
-        )
-
-        assert r2.ping() is True
-
-    @skip_if_redis_enterprise()
-    def test_credentials_provider_with_supplier(self, r, request):
-        import functools
-
-        @functools.lru_cache(maxsize=10)
-        def auth_supplier(user, endpoint):
-            def get_random_string(length):
-                letters = string.ascii_lowercase
-                result_str = "".join(random.choice(letters) for i in range(length))
-                return result_str
-
-            auth_token = get_random_string(5) + user + "_" + endpoint
-            return user, auth_token
-
-        username = "redis-py-auth"
-        creds_provider = CredentialsProvider(
-            supplier=auth_supplier,
-            user=username,
-            endpoint="localhost",
-        )
-        password = creds_provider.get_password()
-
-        assert r.acl_setuser(
-            username,
-            enabled=True,
-            passwords=["+" + password],
-            keys="~*",
-            commands=["+ping", "+command", "+info", "+select", "+flushdb", "+cluster"],
-        )
-
-        def teardown():
-            r.acl_deluser(username)
-
-        request.addfinalizer(teardown)
-
-        r2 = _get_client(
-            redis.Redis, request, flushdb=False, credentials_provider=creds_provider
-        )
-
-        assert r2.ping() is True
